Last fixes - WORKING

+++ src/telnetServerCommands.cpp
@@ -52,7 +52,7 @@ bool send_addCode(Commander &Cmdr)
 {
     int Dst;
     int type;
-    int code;
+    uint32_t code;
     int ButtonBitmap;
     int validStart = 0;
     int validEnd = 0;
@@ -60,7 +60,7 @@ bool send_addCode(Commander &Cmdr)
         goto error;
     if(! Cmdr.getInt(type))
         goto error;
-    if(! Cmdr.getInt(code))
+    if(! Cmdr.getLong(code))
         goto error;
     if(! Cmdr.getInt(ButtonBitmap))
         goto error;
